adding more organized views, api example with crayons data, and documentation forit
