+++ docs/_docs/getting-started/index.md
@@ -12,3 +12,4 @@ The documentation here will get you started to use the GCP Flask Stanford templa
 
  - [Getting Started](development) with development and setup for your needs.
  - [Templates](templates) and examples available for customization.
+ - [Application Programming Interface](api.md) example for a RESTful API
