+++ gcpflask/server.py
@@ -1,5 +1,6 @@
 from flask import Flask
 from flask_login import LoginManager
+from flask_restful import Api
 from flask_wtf.csrf import CSRFProtect, generate_csrf
 from flask_cors import CORS
 from .database import *
@@ -46,6 +47,9 @@ login_manager = LoginManager()
 login_manager.init_app(app)
 login_manager.login_view = "login"
 login_manager.session_protection = "strong"
+
+# Application Programming interface
+api = Api(app)
 
 # Cors
