+++ gcpflask/static/data/crayons.csv
@@ -1,0 +1,120 @@
+"Mahogany","CD4A4A","205,74,74"
+"Fuzzy Wuzzy Brown","CC6666","204,102,102"
+"Chestnut","BC5D58","188,93,88"
+"Red Orange","FF5349","255,83,73"
+"Sunset Orange","FD5E53","253,94,83"
+"Bittersweet","FD7C6E","253,124,110"
+"Melon","FDBCB4","253,188,180"
+"Outrageous Orange","FF6E4A","255,110,74"
+"Vivid Tangerine","FFA089","255,160,137"
+"Burnt Sienna","EA7E5D","234,126,93"
+"Brown","B4674D","180,103,77"
+"Sepia","A5694F","165,105,79"
+"Orange","FF7538","255,117,56"
+"Burnt Orange","FF7F49","255,127,73"
+"Copper","DD9475","221,148,117"
+"Mango Tango","FF8243","255,130,67"
+"Atomic Tangerine","FFA474","255,164,116"
+"Beaver","9F8170","159,129,112"
+"Antique Brass","CD9575","205,149,117"
+"Desert Sand","EFCDB8","239,205,184"
+"Raw Sienna","D68A59","214,138,89"
+"Tumbleweed","DEAA88","222,170,136"
+"Tan","FAA76C","250,167,108"
+"Peach","FFCFAB","255,207,171"
+"Macaroni and Cheese","FFBD88","255,189,136"
+"Apricot","FDD9B5","253,217,181"
+"Neon Carrot","FFA343","255,163,67"
+"Almond","EFDBC5","239,219,197"
+"Yellow Orange","FFB653","255,182,83"
+"Gold","E7C697","231,198,151"
+"Shadow","8A795D","138,121,93"
+"Banana Mania","FAE7B5","250,231,181"
+"Sunglow","FFCF48","255,207,72"
+"Goldenrod","FCD975","252,217,117"
+"Dandelion","FDDB6D","253,219,109"
+"Yellow","FCE883","252,232,131"
+"Green Yellow","F0E891","240,232,145"
+"Spring Green","ECEABE","236,234,190"
+"Olive Green","BAB86C","186,184,108"
+"Laser Lemon","FDFC74","253,252,116"
+"Unmellow Yellow","FDFC74","253,252,116"
+"Canary","FFFF99","255,255,153"
+"Yellow Green","C5E384","197,227,132"
+"Inch Worm","B2EC5D","178,236,93"
+"Asparagus","87A96B","135,169,107"
+"Granny Smith Apple","A8E4A0","168,228,160"
+"Electric Lime","1DF914","29,249,20"
+"Screamin Green","76FF7A","118,255,122"
+"Fern","71BC78","113,188,120"
+"Forest Green","6DAE81","109,174,129"
+"Sea Green","9FE2BF","159,226,191"
+"Green","1CAC78","281,72,120"
+"Mountain Meadow","30BA8F","48,186,143"
+"Shamrock","45CEA2","69,206,162"
+"Jungle Green","3BB08F","59,176,143"
+"Caribbean Green","1CD3A2","28,211,162"
+"Tropical Rain Forest","17806D","231,28,109"
+"Pine Green","158078","211,28,120"
+"Robin Egg Blue","1FCECB","31,206,203"
+"Aquamarine","78DBE2","120,219,226"
+"Turquoise Blue","77DDE7","119,221,231"
+"Sky Blue","80DAEB","128,218,235"
+"Outer Space","414A4C","65,74,76"
+"Blue Green","199EBD","251,58,189"
+"Pacific Blue","1CA9C9","28,169,201"
+"Cerulean","1DACD6","29,172,214"
+"Cornflower","9ACEEB","154,206,235"
+"Midnight Blue","1A4876","26,72,118"
+"Navy Blue","1974D2","25,116,210"
+"Denim","2B6CC4","43,108,196"
+"Blue","1F75FE","31,117,254"
+"Periwinkle","C5D0E6","197,208,230"
+"Cadet Blue","B0B7C6","176,183,198"
+"Indigo","5D76CB","93,118,203"
+"Wild Blue Yonder","A2ADD0","162,173,208"
+"Manatee","979AAA","151,154,170"
+"Blue Bell","ADADD6","173,173,214"
+"Blue Violet","7366BD","115,102,189"
+"Purple Heart","7442C8","116,66,200"
+"Royal Purple","7851A9","120,81,169"
+"Purple Mountains’ Majesty","9D81BA","157,129,186"
+"Violet (Purple)","926EAE","146,110,174"
+"Wisteria","CDA4DE","205,164,222"
+"Vivid Violet","8F509D","143,80,157"
+"Fuchsia","C364C5","195,100,197"
+"Shocking Pink","FB7EFD","251,126,253"
+"Pink Flamingo","FC74FD","252,116,253"
+"Plum","8E4585","142,69,133"
+"Hot Magenta","FF1DCE","255,29,206"
+"Purple Pizzazz","FF1DCE","255,29,206"
+"Razzle Dazzle Rose","FF48D0","255,72,208"
+"Orchid","E6A8D7","230,168,215"
+"Red Violet","C0448F","192,68,143"
+"Eggplant","6E5160","110,81,96"
+"Cerise","DD4492","221,68,146"
+"Wild Strawberry","FF43A4","255,67,164"
+"Magenta","F664AF","246,100,175",
+"Lavender","FCB4D5","252,180,213"
+"Cotton Candy","FFBCD9","255,188,217"
+"Violet Red","F75394","247,83,148"
+"Carnation Pink","FFAACC","255,170,204"
+"Razzmatazz","E3256B","227,37,107"
+"Piggy Pink","FDD7E4","253,215,228"
+"Jazzberry Jam","CA3767","202,55,103"
+"Blush","DE5D83","222,93,131"
+"Tickle Me Pink","FC89AC","252,137,172"
+"Pink Sherbet","F780A1","247,128,161"
+"Maroon","C8385A","200,56,90"
+"Red","EE204D","238,32,77"
+"Radical Red","FF496C","255,73,108"
+"Mauvelous","EF98AA","239,152,170"
+"Wild Watermelon","FC6C85","252,108,133"
+"Scarlet","FC2847","252,40,71"
+"Salmon","FF9BAA","255,155,170"
+"Brick Red","CB4154","203,65,84"
+"White","EDEDED","237,237,237"
+"Timberwolf","DBD7D2","219,215,210"
+"Silver","CDC5C2","205,197,194"
+"Gray","95918C","149,145,140"
+"Black","232323","35,35,35"
