+++ gcpflask/templates/cardinal/base.html
@@ -122,7 +122,7 @@
                      <li><a  href="#">Submenu 5</a></li>
                    </ul>
                   </li>
-                  <li id="nav-7">{% if current_user.is_authenticated %}<a href="/logout">Logout</a>{% else %}<a href="/login">Login</a></li>{% endif %}
+                  <li id="nav-7">{% if current_user.is_authenticated %}<a href="/cardinal/logout">Logout</a>{% else %}<a href="/cardinal/login">Login</a></li>{% endif %}
                 </ul>
               </div>
             </div>
