+++ gcpflask/views/__init__.py
@@ -1,0 +1,3 @@
+from .main import *
+from .api import *
+from .cardinal import *
